Update models.py
The previous message was misleading as it referred to grant types that aren't checked in `default_redirect_uri`. 'Client credentials' grant was not specified here, but the `clean` method was passing earlier on. `ImplicitGrant` is being mistakenly used in the OAuthLib backend, which I'm also investigating.

Also some minor PEP8 fixes and an `AbstractApplication.GRANTS_WITH_REDIRECT` attribute for neater conditionals.